LayoutTests/js/ShadowRealm-iframe-sandboxed-expected.txt

Test to ensure ShadowRealms running in a sandboxed iframe do not access the parent window object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


attached frames: 1
detaching iframe
dropping realm value
PASS () => checkFn2(wrappedLog("shadowRealm")) is true
PASS successfullyParsed is true

TEST COMPLETE
